Move helper function `use_color` to `tools/options.py`.
